An S4 Class that stores a fitted coarse data object

This is the output from dic.fit(), which contains the important bits of information about the model fit and key options used.

Slots

ests:

Matrix of class "numeric". This matrix summarizes the results of fitting the model. Rows correspond to the first parameter , the second parameter and then percentiles specified by the ptiles argument. Columns correspond to the point estimate, the lower and upper bounds on the 95% confidence interval and the standard error of the point estimate. If the maximization does not converge, this matrix is filled with NAs.

conv:

Object of class "numeric". A value of 1 indicates successful convergence; 0 indicates unsuccessful convergence.

MSG:

Object of class "character". The error message returned from optim() if the routine fails to converge.

loglik:

Object of class "numeric". Value of the estimated maximum log-likelihood.

samples:

Object of class "data.frame". Data frame of bootstrap estimates of parameters (if bootstraps were performed).

data:

Object of class "data.frame". Original data used to fit model.

dist:

Object of class "character". Failure time distribution fit to data. "L" for log-normal, "G" for gamma, "W" for Weibull, and "E" for Erlang.

inv.hessian:

Object of class "matrix". The inverse of the hessian matrix for the likelihood surface at the MLE. Used to determine the standard errors for the percentiles. Note that optimization is done on a transformed scale with all parameters logged for all distributions except the first parameter of the log-normal distribution.

est.method:

Object of class "character". Method used for estimation.

ci.method:

Object of class "character". Method used for estimation of confidence/credible intervals.