cd.fit.mcmc {coarseDataTools}	R Documentation

An S4 Class that stores a MCMC fit coarse data object

Description

This is the output from dic.fit.mcmc(), which contains the important bits of information about the model fit and key options used.

Slots

ests:

Matrix of class "numeric". This matrix summarizes the results of fitting the model. Rows correspond to the first parameter , the second parameter and then percentiles specified by the ptiles argument. Columns correspond to the point estimate, the lower and upper bounds on the 95% credible interval and the standard error of the point estimate.

conv:

Object of class "numeric". Not used in with dic.fit.mcmc

MSG:

Object of class "character". The error message returned from optim() if the routine fails to converge.

loglik:

Object of class "numeric". Not used in with dic.fit.mcmc.

samples:

Object of class "data.frame". Data frame of posterior draws of parameters.

data:

Object of class "data.frame". Original data used to fit model.

dist:

Object of class "character". Failure time distribution fit to data. "L" for log-normal, "G" for gamma, "W" for Weibull, and "E" for Erlang.

inv.hessian:

Object of class "matrix". Not used in with dic.fit.mcmc.

est.method:

Object of class "character". Method used for estimation.

ci.method:

Object of class "character". Method used for estimation of confidence/credible intervals.

[Package coarseDataTools version 0.6-5 Index]