dgammaOff1 {coarseDataTools}	R Documentation

Function that calculates dgamma with a offset of 1 (i.e., 1 is equivalent to 0)

Description

Function that calculates dgamma with a offset of 1 (i.e., 1 is equivalent to 0)

Usage

dgammaOff1(x, replace0 = FALSE, ...)

Arguments

x
value to calculate dgamma at

replace0
should we replace 0 with epsilon

...
other parameters to dgamma

Value

dgamma offset

[Package coarseDataTools version 0.6-5 Index]