exp.win.lengths {coarseDataTools}	R Documentation

Exposure window lengths from an influenza outbreak at a NYC school

Description

A numeric vector of exposure window lengths taken from a dataset of doubly interval-censored incubation period observations. All observations came from a NYC public school. The outbreak has been described in full in Lessler et al. (see citation below).

Usage

data(exp.win.lengths)

Format

A numeric vector with 134 positive values. Each value represents an exposure window length from an observation of the incubation period for that individual. The exposure window length is the length of time during which exposure could have occurred. For example, if an individual could have been exposed anytime between 6am on Monday to 6am on Wednesday, her exposure window length would be 2 days.

Source

Lessler J et al. New England Journal of Medicine. Outbreak of 2009 Pandemic Influenza A (H1N1) at a New York City School. 2009. 361(27):2628-2636. http://content.nejm.org/cgi/content/full/361/27/2628

Examples

data(exp.win.lengths)
summary(exp.win.lengths)
hist(exp.win.lengths)

[Package coarseDataTools version 0.6-5 Index]