Coarse incubation period data for influenza A

These observations on the incubation period of influenza A come from a variety of sources, and were gathered for a literature review. They report doubly interval-censored, single interval-censored or exact observations for the incubation period.

Usage

data(fluA.inc.per)

Format

A data frame with 151 observations on the following 7 variables.

author

the name of the primary author for the source of the observation

year

the year of the study which is the source of the observation

EL

the earliest possible time of infection

ER

the latest possible time of infection

SL

the earliest possible time of symptom onset

SR

the latest possible time of symptom onset

type

an indicator of the type of observation: 0 for doubly interval-censored, 1 for single-interval censored, 2 for exact

Source

Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT. (2009) A systematic review of the incubation periods of acute respiratory viral infections. Lancet Infectious Diseases. 9(5):291-300.

Examples

data(fluA.inc.per)
head(fluA.inc.per)
#>   author year EL  ER  SL  SR type
#> 1  moser 1977  0 0.5 1.0 1.5    0
#> 2  moser 1977  0 0.5 1.5 2.0    0
#> 3  moser 1977  0 0.5 1.5 2.0    0
#> 4  moser 1977  0 0.5 1.5 2.0    0
#> 5  moser 1977  0 0.5 1.5 2.0    0
#> 6  moser 1977  0 0.5 1.5 2.0    0