Negative log likelihood for a dataset of interval-censored data, given a distribution and its parameters.

Negative log likelihood for a dataset of interval-censored data, given a distribution and its parameters.

Usage

loglikhd(pars, dat, dist)

Arguments

pars

vector of the transformed (estimation scale) parameters

dat

a dataset, as in dic.fit

dist

a distribution, as in dic.fit

Value

negative log-likelihood for a given dataset, parameters, and distribution.

Details

This package uses two versions of each parameter, the estimation scale, or the scale that is used for numerical optimization, and the reporting scale, or the natural scale of the parameters. For all likelihood calculations, this loglikhd function expects parameters that are on the estimation scale, i.e. have range \((-\infty, \infty)\). Specifically, this translates into all parameters for all distributions being log-transformed except for the meanlog (i.e. "par1") for the log-normal distribution.