Does a metropolis hastings for the Erlang distribution

Does a metropolis hastings for the Erlang distribution

Usage

mcmc.erlang(
  dat,
  prior.par1,
  prior.par2,
  init.pars,
  verbose,
  burnin,
  n.samples,
  sds = c(1, 1)
)

Arguments

dat

the data to fit

prior.par1

mean of priors. A negative binomial (for shape) and a normal for log(scale)

prior.par2

dispersion parameters for priors, dispersion for negative binomial, log scale sd for normal

init.pars

the starting parameters on the reporting scale

verbose

how often to print an update

burnin

how many burnin iterations to do

n.samples

the number of samples to keep and report back

sds

the standard deviations for the proposal distribution

Value

a matrix of n.samples X 2 parameters, on the estimation scale