nycH1N1 {coarseDataTools}	R Documentation

Incubation period data from New York City Public Schools, 2009 H1N1 influenza outbreak

Description

These observations on the incubation period of influenza A come from the investigation of the H1N1 outbreak in NYC schools in the spring of 2009. They report doubly interval-censored observations for the incubation period.

Usage

data(nycH1N1)

Format

A data frame with 134 observations on the following 5 variables.

EL

the earliest possible time of infection

ER

the latest possible time of infection

SL

the earliest possible time of symptom onset

SR

the latest possible time of symptom onset

type

an indicator of the type of observation: 0 for doubly interval-censored, 1 for single-interval censored, 2 for exact. All of these observations are doubly interval-censored.

Source

Lessler J, Reich NG, Cummings DAT and The DOHMH Swine Influenza Investigation Team. Outbreak of 2009 Pandemic Influenza A (H1N1) at a New York City School. New England Journal of Medicine. 2009. 361(27):2628-2636.

Examples

data(nycH1N1)
head(nycH1N1)
## Not run: dic.fit(nycH1N1)

[Package coarseDataTools version 0.6-5 Index]