plot,cd.fit-method {coarseDataTools}	R Documentation

Plots the estimated survival function with an option to plot the posterior draws or bootstraps behind the fit

Description

Plots the estimated survival function with an option to plot the posterior draws or bootstraps behind the fit

Usage

## S4 method for signature 'cd.fit'
plot(x, y, col.main = rgb(230, 85, 13, maxColorValue =
  255), col.samps = rgb(99, 99, 99, 10, maxColorValue = 255),
  plot.n.samps = 200, add = F, xlim = range(x@data), ...)

Arguments

x
cd.fit or cd.fit.mcmc object

y
unused

col.main
color for plotting the main estimate

col.samps
color for the samples (should include some alpha transparency)

plot.n.samps
how many posterior or bootstrap samples do you want to plot?

add
add to existing plot?

xlim
xlim for plot, defaults to the range of the data slot in x

...
other options to pass to plot

[Package coarseDataTools version 0.6-5 Index]