simulated.outbreak.deaths {coarseDataTools}	R Documentation

Simulated case and death reports from a fictional outbreak

Description

This dataset provides reported counts of cases and deaths occurring at different time points across a simulated outbreak. Details of the data simulation algorithm are provided in the manuscript "Estimating case fatality ratios from infectious disease surveillance data" (Reich et al., under review, available upon request).

Usage

data(simulated.outbreak.deaths)

Format

time

time, t, after start of outbreak

grp

an categorical variable indicating membership in one of two groups of a covariate, j

R

number of recovered cases reported at the given t and j

D

number of deaths reported at the given t and j

N

total number of cases and deaths reported at t and j, or D+R

Source

Reich NG, Lessler J, Cummings DAT, Brookmeyer R. Estimating case fatality ratios from infectious disease surveillance data. Biometrics. 2012. 68(2): 598-606.

Examples

data(simulated.outbreak.deaths)
head(simulated.outbreak.deaths)
plot(simulated.outbreak.deaths[simulated.outbreak.deaths[,"grp"]==1,"D"], type="l")

[Package coarseDataTools version 0.6-5 Index]